+++ apps/ymir-sdl3/src/app/shared_context.hpp
@@ -23,10 +23,10 @@
 
 #include <blockingconcurrentqueue.h>
 
+#include <deque>
 #include <filesystem>
 #include <memory>
 #include <mutex>
-#include <deque>
 
 namespace app {